# docs/openapi.yaml
openapi: 3.1.0
info:
  title: meili-ingest Gateway API
  version: 0.1.0
  summary: Ingest any content into Meilisearch through configurable pipelines.
  description: |
    The gateway is the public entry point of meili-ingest.

    **Two tiers**

    * `POST /ingest` — auto-routing. The gateway detects the MIME type (magic
      bytes → file extension → UTF-8 sniff), selects the best matching pipeline
      (tenant-scoped user pipelines > global user pipelines > built-ins),
      resolves the target index and starts a Temporal workflow.
    * `POST /ingest/pipeline/{name}` — explicit pipeline, full control.

    **Tenant context**

    Every request is scoped to a `MeiliContext` resolved from, in order:
    `X-Meili-Host` / `X-Meili-Api-Key` / `X-Meili-Project-Id` / `X-Meili-Index`
    headers (injected by Meilisearch Cloud's Envoy and trusted only when
    `X-Meili-Envoy-Secret` matches the gateway's `ENVOY_TRUSTED_HEADER`), the
    `?index=` query parameter, an `Authorization: Bearer <key>` header, and
    finally the gateway's `MEILI_URL` / `MEILI_API_KEY` environment variables.
    When none of these yields a host and API key the request fails with `400` —
    unless the pipeline is **pinned**: every `meili_indexer` step names a
    [Meilisearch connection](/concepts/connections), which then supplies the host
    and key, and no context is needed.

    **Errors** are always `{"error": "<message>", "code": "<snake_case>"}`.
  license:
    name: MIT
...
  contact:
...
servers:
...
security:
  - bearerAuth: []
  - {}
...
paths:
...
  /connections:
    get:
      tags: [Connections]
      operationId: listConnections
      summary: List Meilisearch connections
      description: The caller's project's connections, then global ones. Keys are always masked.
      parameters:
        - $ref: "#/components/parameters/XMeiliProjectId"
        - $ref: "#/components/parameters/XMeiliEnvoySecret"
      responses:
        "200":
          description: Connections.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Connection"
        "501":
          $ref: "#/components/responses/NotEnabled"
        "502":
          $ref: "#/components/responses/UpstreamUnavailable"
    post:
      tags: [Connections]
      operationId: createConnection
      summary: Create a connection
      description: |
        Checks `host` against `MEILI_CONNECTION_HOSTS`, then calls `GET {host}/health`
        and `GET {host}/indexes?limit=1` with the key. Any failure is `422` and
        nothing is stored. The key is sealed before it leaves the gateway.
      parameters:
        - $ref: "#/components/parameters/XMeiliProjectId"
        - $ref: "#/components/parameters/XMeiliEnvoySecret"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConnection"
            example:
              uid: prod
              name: Production cluster
              host: https://ms-1234.meilisearch.io
              api_key: "<write key>"
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "501":
          $ref: "#/components/responses/NotEnabled"
        "502":
          $ref: "#/components/responses/UpstreamUnavailable"
  /connections/{uid}:
    parameters:
      - $ref: "#/components/parameters/ConnectionUid"
      - $ref: "#/components/parameters/XMeiliProjectId"
      - $ref: "#/components/parameters/XMeiliEnvoySecret"
    get:
      tags: [Connections]
      operationId: getConnection
      summary: Get a connection
      responses:
        "200":
          description: The connection, with `used_by`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "404":
          $ref: "#/components/responses/NotFound"
        "501":
          $ref: "#/components/responses/NotEnabled"
    patch:
      tags: [Connections]
      operationId: updateConnection
      summary: Update a connection
      description: Changing `host` or `api_key` re-runs the create-time checks. Omit `api_key` to keep the stored key.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateConnection"
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "501":
          $ref: "#/components/responses/NotEnabled"
    delete:
      tags: [Connections]
      operationId: deleteConnection
      summary: Delete a connection
      description: Never blocked. Pipelines that still name it fail at run time with `connection "<uid>" not found`.
      responses:
        "204":
          description: Deleted.
        "404":
          $ref: "#/components/responses/NotFound"
        "501":
          $ref: "#/components/responses/NotEnabled"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Self-hosted fallback: a Meilisearch API key with write access. Used as
        `api_key` when no `X-Meili-Api-Key` header is present. Combine with
        `X-Meili-Host` (or the gateway's `MEILI_URL`) to target an instance.
  parameters:
...
    ConnectionUid:
      name: uid
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-zA-Z0-9._-]+$"
      example: prod
...
    XMeiliProjectId:
      name: X-Meili-Project-Id
      in: header
      required: false
      description: Tenant / project identifier (Envoy-injected).
      schema:
        type: string
      example: xxx
...
    XMeiliEnvoySecret:
      name: X-Meili-Envoy-Secret
      in: header
      required: false
      description: |
        Shared secret proving the request came through Envoy. Required for the
        other `X-Meili-*` headers to be honoured when the gateway runs with
        `ENVOY_TRUSTED_HEADER` set; ignored otherwise.
      schema:
        type: string
  responses:
...
    BadRequest:
      description: Malformed request or missing tenant context (no `X-Meili-*` headers, no bearer key and no `MEILI_URL`).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { "error": "no Meilisearch context: set X-Meili-Host/X-Meili-Api-Key headers or MEILI_URL/MEILI_API_KEY", "code": "missing_context" }
...
    NotFound:
      description: Unknown connection or source.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { "error": "source \"tmdb\" not found", "code": "not_found" }
    Unprocessable:
      description: Validation failed (bad cron or template, forbidden host — including an ingest `url` outside `SOURCE_FETCH_HOSTS` — unreachable Meilisearch or rejected key, unpinned pipeline).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { "error": "pipeline \"movies\" cannot run on a schedule: its meili_indexer step names no Meilisearch `connection`", "code": "validation" }
    NotEnabled:
      description: "`SOURCE_SECRET_KEY` is not set on this deployment, so connections and sources are disabled."
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UpstreamUnavailable:
      description: Control plane or Temporal unreachable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example: { "error": "control plane unreachable", "code": "upstream_error" }
  schemas:
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable message.
        code:
          type: string
          description: Stable snake_case error code.
          examples: [missing_context, bad_request, forbidden, not_found, payload_too_large, unsupported_media_type, invalid_pipeline, upstream_error, internal_error]
...
    Connection:
      type: object
      required: [uid, name, host, api_key, created_at, updated_at]
      properties:
        uid:
          type: string
          description: Name a `meili_indexer` step references in `config.connection`.
        name:
          type: string
        tenant_id:
          type: string
          description: Tenant scope. Absent = global.
        host:
          type: string
          format: uri
        api_key:
          type: string
          const: "****"
          description: Always masked.
        used_by:
          type: array
          items: { type: string }
          description: Pipelines whose indexer names this connection (single reads only).
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CreateConnection:
      type: object
      required: [uid, host, api_key]
      additionalProperties: false
      properties:
        uid:
          type: string
          pattern: "^[a-zA-Z0-9._-]+$"
        name:
          type: string
          description: Defaults to `uid`.
        host:
          type: string
          format: uri
        api_key:
          type: string
          writeOnly: true
    UpdateConnection:
      type: object
      additionalProperties: false
      properties:
        name: { type: string }
        host: { type: string, format: uri }
        api_key:
          type: string
          writeOnly: true
          description: Omit to keep the stored key.
...